Updated return of the validate_model_parameter_file
With the old return values, the only way to check if validation passed was to
check if the returned value was TRUE. However, when validation failed and the
function returned the list of errors, this check would generate a warning from
R "the condition has length > 1 and only the first element will be used". With
this change, users can check the `success` field for TRUE to see if validation
passed, and since its never a vector, the above warning will never be
generated.